requirements.txt
streamlit
opencv-python-headless
numpy
imutils
Pillow